Added some security to MoviesController.cs functions

+++ MvcVeevonauts/MvcVeevonauts/Controllers/MoviesController.cs
@@ -24,9 +24,13 @@ namespace MvcVeevonauts.Controllers
         //
         // GET: /Movies/Details/5
 
-        public ViewResult Details(int id)
+        public ActionResult Details(int id)
         {
             Movie movie = db.Movies.Find(id);
+            if (movie == null)
+            {
+                return HttpNotFound();
+            }
             return View(movie);
         }
 
@@ -85,9 +89,13 @@ namespace MvcVeevonauts.Controllers
         //
         // GET: /Movies/Delete/5
  
-        public ActionResult Delete(int id)
+        public ActionResult Delete(int id = 0)
         {
             Movie movie = db.Movies.Find(id);
+            if (movie == null)
+            {
+                return HttpNotFound();
+            }
             return View(movie);
         }
 
@@ -95,9 +103,13 @@ namespace MvcVeevonauts.Controllers
         // POST: /Movies/Delete/5
 
         [HttpPost, ActionName("Delete")]
-        public ActionResult DeleteConfirmed(int id)
+        public ActionResult DeleteConfirmed(int id = 0)
         {            
             Movie movie = db.Movies.Find(id);
+            if (movie == null)
+            {
+                return HttpNotFound();
+            }
             db.Movies.Remove(movie);
             db.SaveChanges();
             return RedirectToAction("Index");
